extracted utility functions into separate header
